+++ packages/backend-proxy-middleware/src/base/proxy.ts
@@ -15,6 +15,7 @@ import {
     isFullUrlDestination,
     BAS_DEST_INSTANCE_CRED_HEADER
 } from '@sap-ux/btp-utils';
+import type { ServiceInfo } from '@sap-ux/btp-utils';
 import type { BackendConfig, DestinationBackendConfig, LocalBackendConfig } from './types';
 import translations from './i18n.json';
 
@@ -22,6 +23,7 @@ import type { ApiHubSettings, ApiHubSettingsKey, ApiHubSettingsService, BackendS
 import { AuthenticationType, BackendSystemKey, getService } from '@sap-ux/store';
 import { getCorporateProxyServer, isHostExcludedFromProxy } from './config';
 import type { Url } from 'url';
+import { addOptionsForEmbeddedBSP } from '../ext/bsp';
 
 /**
  * Collection of custom event handler for the proxy.
@@ -239,7 +241,7 @@ export async function enhanceConfigForSystem(
         if (system.serviceKeys) {
             const provider = createForAbapOnCloud({
                 environment: AbapCloudEnvironment.Standalone,
-                service: JSON.parse(system.serviceKeys as string),
+                service: system.serviceKeys as ServiceInfo,
                 refreshToken: system.refreshToken,
                 refreshTokenChangedCb: tokenChangedCallback
             });
@@ -257,13 +259,8 @@ export async function enhanceConfigForSystem(
         // sending a request to the backend to get cookies
         await provider.getAtoInfo();
         proxyOptions.headers['cookie'] = provider.cookies.toString();
-    } else if (
-        (system.username || process.env.FIORI_TOOLS_USER) &&
-        (system.password || process.env.FIORI_TOOLS_PASSWORD)
-    ) {
-        proxyOptions.auth = `${system.username || process.env.FIORI_TOOLS_USER}:${
-            system.password || process.env.FIORI_TOOLS_PASSWORD
-        }`;
+    } else if (system.username && system.password) {
+        proxyOptions.auth = `${system.username}:${system.password}`;
     }
 }
 
@@ -315,7 +312,15 @@ export async function generateProxyMiddlewareOptions(
         }
     }
 
+    if (!proxyOptions.auth && process.env.FIORI_TOOLS_USER && process.env.FIORI_TOOLS_PASSWORD) {
+        proxyOptions.auth = `${process.env.FIORI_TOOLS_USER}:${process.env.FIORI_TOOLS_PASSWORD}`;
+    }
+
     proxyOptions.pathRewrite = PathRewriters.getPathRewrite(backend, logger);
+
+    if (backend.bsp) {
+        await addOptionsForEmbeddedBSP(backend.bsp, proxyOptions, logger);
+    }
 
     if (backend.apiHub) {
         const apiHubKey = await getApiHubKey(logger);