+++ packages/fe-fpm-writer/test/integration/index.test.ts
@@ -6,9 +6,11 @@ import {
     generateCustomColumn,
     generateCustomPage,
     TargetControl,
-    generateCustomSection
+    generateCustomSection,
+    generateCustomView
 } from '../../src';
 import { Placement } from '../../src/common/types';
+import { generateListReport, generateObjectPage } from '../../src/page';
 
 describe('use FPM with existing apps', () => {
     const testInput = join(__dirname, '../test-input');
@@ -26,10 +28,30 @@ describe('use FPM with existing apps', () => {
         }
     });
 
-    describe('extend Fiori elements for OData v4 ListReport ObjectPage', () => {
+    describe('extend UI5 application with FPM', () => {
         const targetPath = join(testOutput, 'lrop');
+        const mainEntity = 'Travel';
         beforeAll(() => {
             fs.copy(join(testInput, 'basic-lrop'), targetPath);
+        });
+
+        test('generateListReport', () => {
+            generateListReport(targetPath, { entity: mainEntity }, fs);
+        });
+
+        test('generateObjectPage with navigation from ListReport', () => {
+            generateObjectPage(
+                targetPath,
+                {
+                    entity: mainEntity,
+                    navigation: {
+                        navEntity: mainEntity,
+                        sourcePage: 'TravelListReport',
+                        navKey: true
+                    }
+                },
+                fs
+            );
         });
 
         test('generateCustomPage with navigation from ObjectPage', () => {
@@ -39,7 +61,6 @@ describe('use FPM with existing apps', () => {
                     name: 'MyCustomPage',
                     entity: 'Booking',
                     navigation: {
-                        sourceEntity: 'Travel',
                         sourcePage: 'TravelObjectPage',
                         navEntity: '_Booking'
                     }
@@ -52,7 +73,7 @@ describe('use FPM with existing apps', () => {
             generateCustomColumn(
                 targetPath,
                 {
-                    target: 'TravelList',
+                    target: 'TravelListReport',
                     targetEntity: '@com.sap.vocabularies.UI.v1.LineItem',
                     name: 'NewCustomColumn',
                     header: 'Custom Price and Currency',
@@ -67,19 +88,56 @@ describe('use FPM with existing apps', () => {
             );
         });
 
+        test('generateCustomView in ListReport', () => {
+            //pre-requisite is at least one view based on annotations
+            fs.extendJSON(join(targetPath, 'webapp/manifest.json'), {
+                'sap.ui5': {
+                    routing: {
+                        targets: {
+                            TravelListReport: {
+                                options: {
+                                    settings: {
+                                        views: {
+                                            paths: [
+                                                {
+                                                    key: 'LineItemView',
+                                                    annotationPath: 'com.sap.vocabularies.UI.v1.LineItem'
+                                                }
+                                            ]
+                                        }
+                                    }
+                                }
+                            }
+                        }
+                    }
+                }
+            });
+            generateCustomView(
+                targetPath,
+                {
+                    target: 'TravelListReport',
+                    key: 'CustomViewKey',
+                    label: 'Custom View',
+                    name: 'NewCustomView',
+                    eventHandler: true
+                },
+                fs
+            );
+        });
+
         test('generateCustomAction in ListReport and ObjectPage', () => {
             generateCustomAction(
                 targetPath,
                 {
                     name: 'MyCustomAction',
                     target: {
-                        page: 'TravelList',
+                        page: 'TravelListReport',
                         control: TargetControl.table
                     },
                     settings: {
-                        text: 'My Custom Action',
-                        eventHandler: true
-                    }
+                        text: 'My Custom Action'
+                    },
+                    eventHandler: true
                 },
                 fs
             );
@@ -92,9 +150,36 @@ describe('use FPM with existing apps', () => {
                         control: TargetControl.header
                     },
                     settings: {
-                        text: 'My other Action',
-                        eventHandler: true
-                    }
+                        text: 'My other Action'
+                    },
+                    eventHandler: true
+                },
+                fs
+            );
+            // Generate custom action by appending existing file
+            generateCustomAction(
+                targetPath,
+                {
+                    name: 'AppendedAction',
+                    target: {
+                        page: 'TravelObjectPage',
+                        control: TargetControl.header
+                    },
+                    settings: {
+                        text: 'My other Action'
+                    },
+                    eventHandler: {
+                        fileName: 'AnotherCustomAction',
+                        fnName: 'OnAppendedFn',
+                        insertScript: {
+                            fragment: `,\n        OnAppendedFn: function() {\n            MessageToast.show("Custom handler invoked.");\n        }`,
+                            position: {
+                                line: 8,
+                                character: 9
+                            }
+                        }
+                    },
+                    folder: join('ext', 'anotherCustomAction')
                 },
                 fs
             );
@@ -118,11 +203,7 @@ describe('use FPM with existing apps', () => {
         });
 
         afterAll(() => {
-            if (!debug) {
-                expect(
-                    (fs as any).dump(relative(process.cwd(), targetPath), '**/webapp/{manifest.json,ext/**/*}')
-                ).toMatchSnapshot();
-            }
+            expect((fs as any).dump(targetPath, '**/test-output/*/webapp/{manifest.json,ext/**/*}')).toMatchSnapshot();
         });
     });
 });