+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,63 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.6.6
+
+### Patch Changes
+
+-   cc1c406: chore(open-ux-tools) ignore source map files when publishing to npm
+-   Updated dependencies [cc1c406]
+    -   @sap-ux/axios-extension@0.7.2
+    -   @sap-ux/btp-utils@0.10.2
+    -   @sap-ux/logger@0.2.1
+    -   @sap-ux/store@0.3.1
+
+## 0.6.5
+
+### Patch Changes
+
+-   6fae741: Support setting credentials via .env file
+
+## 0.6.4
+
+### Patch Changes
+
+-   febef3e: Moves bsp logic to function generateProxyMiddlewareOptions
+
+## 0.6.3
+
+### Patch Changes
+
+-   Updated dependencies [ebc59b4]
+    -   @sap-ux/store@0.3.0
+
+## 0.6.2
+
+### Patch Changes
+
+-   Updated dependencies [6f0f217]
+    -   @sap-ux/btp-utils@0.10.1
+    -   @sap-ux/axios-extension@0.7.1
+
+## 0.6.1
+
+### Patch Changes
+
+-   5c5c904: Add author to package.json
+
+## 0.6.0
+
+### Minor Changes
+
+-   6f51973: chore(open-ux-tools) Remove node 12 from the list of supported engines for all modules
+
+### Patch Changes
+
+-   Updated dependencies [6f51973]
+    -   @sap-ux/axios-extension@0.7.0
+    -   @sap-ux/btp-utils@0.10.0
+    -   @sap-ux/logger@0.2.0
+    -   @sap-ux/store@0.2.0
+
 ## 0.5.2
 
 ### Patch Changes