+++ packages/btp-utils/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/btp-utils
 
+## 0.10.2
+
+### Patch Changes
+
+-   cc1c406: chore(open-ux-tools) ignore source map files when publishing to npm
+
+## 0.10.1
+
+### Patch Changes
+
+-   6f0f217: Fix getting credentials for a destination service
+
+## 0.10.0
+
+### Minor Changes
+
+-   6f51973: chore(open-ux-tools) Remove node 12 from the list of supported engines for all modules
+
 ## 0.9.2
 
 ### Patch Changes