+++ packages/odata-service-writer/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @sap-ux/odata-service-writer
 
+## 0.14.1
+
+### Patch Changes
+
+-   cc1c406: chore(open-ux-tools) ignore source map files when publishing to npm
+-   Updated dependencies [cc1c406]
+    -   @sap-ux/ui5-config@0.14.1
+
+## 0.14.0
+
+### Minor Changes
+
+-   c29bf99: Support for different project structures
+
+## 0.13.0
+
+### Minor Changes
+
+-   6f51973: chore(open-ux-tools) Remove node 12 from the list of supported engines for all modules
+
+### Patch Changes
+
+-   Updated dependencies [6f51973]
+    -   @sap-ux/ui5-config@0.14.0
+
 ## 0.12.2
 
 ### Patch Changes